Allow disabling editing part of the mappings
For now, this is done over CLI options, it cannot be changed through the
GUI. The new CLI options are:

    -C, --no-edit-classes   Disable editing class names
    -D, --no-edit-javadocs  Disable editing Javadocs
    -E, --no-edit-all       Disable editing everything
    -F, --no-edit-fields    Disable editing field names
    -L, --no-edit-locals    Disable editing local variable and parameter names
    -M, --no-edit-methods   Disable editing method names
    -c, --edit-classes      Enable editing class names
    -d, --edit-javadocs     Enable editing Javadocs
    -e, --edit-all          Enable editing everything
    -f, --edit-fields       Enable editing field names
    -l, --edit-locals       Enable editing local variable and parameter names
    -m, --edit-methods      Enable editing method names

They can be combined together and are applied in order of appearance in the
argument list from left to right, and everything is editable by default,
as before. For example,

    --no-edit-all --edit-javadocs --edit-locals

(or -Edl for short) will allow you to edit only Javadocs and parameter names.